+++ test/root_finding_tests/wilkinson.jl
@@ -45,9 +45,8 @@ end
 
 # Generate Wilkinson functions W₃ etc.:
 for n in 1:15
-    fn_name = symbol(string("W", subscriptify(n)))
+    fn_name = Symbol("W", subscriptify(n))
     expr = generate_wilkinson_horner(n)
 
     @eval $(fn_name)(x) = $(expr)
 end
-